Adds @flourish/layout module
We created an NPM module that makes it easy to add a header and
footer to templates. There's also a handy `setHeight` function
that lets you change the height of the whole project based on
the visualisations height (or whatever you pass in as a value).
The layout module adds html to a target element (document.body in
this case) adds different sections to that element (header, controls
legend, primary and footer). The contents and styles for these
sections are being stored in `state.layout`. The settings for the
layout are being imported using the import property in the
template.yml.

Because the main visualisation part (#network) should go in the
primary container of Flourish layout. I removed it from the html
and add the container dynamically in the draw function.