Highlight visualisation based on preserved state
The code for highlighting countries in this visualisation was
very closely tied to click events. However, in the Flourish story
player, we'd want to highlight certain countries by switch slides.

I changed the logic in this code so that the highlighting part is
more abstract, and is called every time "update()" is called and
there is a currently selected country, not just when an element is
clicked.

I also changed the code so that the current mode is stored in
state, rather than the network object.